% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullreg.R
\name{nullreg}
\alias{nullreg}
\title{Title}
\usage{
nullreg(
  X,
  Y,
  nburn,
  nthin,
  n_save,
  prior.mean.beta = NULL,
  prior.var.beta = NULL,
  prior.sig2 = NULL,
  sig2.samp = NULL,
  beta.samp = NULL,
  verbose = TRUE
)
}
\arguments{
\item{X}{Design matrix}

\item{Y}{Respones variable}

\item{nburn}{MCMC Burnin}

\item{nthin}{MCMC Thinning}

\item{n_save}{MCMC Samples saved}

\item{prior.mean.beta}{Prior mean of the regression coefficient}

\item{prior.var.beta}{Prior variance of the regression coefficient}

\item{prior.sig2}{Prior for the variance term}

\item{sig2.samp}{Starting value for sigma^2}

\item{beta.samp}{Starting value for beta}

\item{verbose}{Print progress}
}
\value{
\preformatted{           Posterior beta and sigma samples
}
}
\description{
Title
}
\examples{
# AR-1 model:
set.seed(1234)
n     = 100
alpha = 0
beta  = 0.9
#beta  = 1.2
sig   = 1
z     = rep(0,n)
z[1]  = alpha/(1-beta)
e     = rnorm(n,0,sig)
for (t in 2:n)
  z[t] = alpha+beta*z[t-1]+e[t]

par(mfrow=c(2,2))
ts.plot(z,main="Simulated AR(1) data")

Y = z[2:n]
x <- z[1:(n-1)]
X <- matrix(x,length(x),1)
nullreg(
X,Y,
nburn = 0 ,nthin = 1,n_save = 1e3,
prior.mean.beta=NULL, prior.var.beta=NULL, prior.sig2=NULL,
sig2.samp= NULL, beta.samp = NULL,
verbose=TRUE
)

}
