% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullreg.R
\name{nullreg}
\alias{nullreg}
\title{Title}
\usage{
nullreg(
  X,
  Y,
  nburn,
  nthin,
  n_save,
  prior.mean.beta = NULL,
  prior.var.beta = NULL,
  prior.sig2 = NULL,
  sig2.samp = NULL,
  beta.samp = NULL,
  intercept = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{X}{Design matrix}

\item{Y}{Respones variable}

\item{nburn}{MCMC Burnin}

\item{nthin}{MCMC Thinning}

\item{n_save}{MCMC Samples saved}

\item{prior.mean.beta}{Prior mean of the regression coefficient}

\item{prior.var.beta}{Prior variance of the regression coefficient}

\item{prior.sig2}{Prior for the variance term}

\item{sig2.samp}{Starting value for sigma^2}

\item{beta.samp}{Starting value for beta}

\item{intercept}{Is an intercept included in the design matrix X}

\item{verbose}{Print progress}
}
\value{
\preformatted{           Posterior beta and sigma samples
}
}
\description{
Title
}
